feat: refactoring hooks and updating function calls in login and register page
