--- frontend/src/hooks/useLogin.ts
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-export const LoginUser = (data: object) => {
-    
-  const url = "http://localhost:8000/session/login";
-  return axios.post(url, data);
-    
-};
