--- frontend/src/hooks/userRegister.ts
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-export const CreateUser = (username: string, data: object) => {
-    
-  const url = `http://localhost:8000/user/create/${username}`;
-  return axios.post(url, data);
-    
-};
